requirements.txt
osmnx>=1.1.1
networkx>=2.5
requests>=2.25.1
shapely>=1.7.1
pandas==2.0.3
matplotlib>=3.3.4
folium>=0.12.1
geopandas==0.13.2
python-dotenv==1.0.0
numpy>=1.19.5
tkintermapview==1.29
scikit-learn>=1.0.2
customtkinter>=5.2.0